# backend/corona-server/src/main/java/de/govhackathon/wvsvcoronatracker/model/User.java
package de.govhackathon.wvsvcoronatracker.model;

import lombok.*;
import org.hibernate.annotations.GenericGenerator;

import javax.persistence.*;
import javax.validation.constraints.NotNull;
import java.util.ArrayList;
import java.util.HashSet;
import java.util.List;
import java.util.Set;


@Data
@EqualsAndHashCode(callSuper = false)
@Builder
@Entity
// quote user, see https://stackoverflow.com/questions/3608420/hibernate-saving-user-model-to-postgres
@Table(name = "\"USER\"")
@AllArgsConstructor
@NoArgsConstructor
public class User {

    @Id
    @GeneratedValue(generator = "UUID")
    @GenericGenerator(
            name = "UUID",
            strategy = "org.hibernate.id.UUIDGenerator"
    )
    private String id;

    @NotNull
    private String token;

    @NotNull
    private String phoneHash;

    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
    private List<HealthDataSet> healthDataSetList = new ArrayList<>();


    @ManyToMany
    @JoinTable(name = "user_friend",
            joinColumns={@JoinColumn(name="user_id", referencedColumnName="id")},
            inverseJoinColumns={@JoinColumn(name="friend_id", referencedColumnName="id")}
    )
    private Set<User> users;

    @Builder.Default
    @ManyToMany(mappedBy = "users", cascade = CascadeType.PERSIST)
    private Set<User> friends = new HashSet<>();


}